Try harder to find bcftools
The bcftools binary is search for with exec.LookPath; however, because
of PBS Pro it is often not on the PATH. As an alternative, we now get
the directory of the current executable (i.e., tso) and search in the
same directory. This is required because generating the excel does not
happen in a container - putting tso in a container comes with its own
problems.